requirements.txt
aiohttp==3.5.4
aiolinebot==0.2
altgraph==0.16.1
async-timeout==3.0.1
attrs==19.1.0
auto-py-to-exe==2.6.5
beautifulsoup4==4.9.0
bottle==0.12.17
bottle-websocket==0.2.9
bs4==0.0.1
cachetools==4.2.2
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
ChatterBot==1.0.8
chatterbot-corpus==1.2.0
Click==7.0
cmake==3.14.4
cymem==2.0.5
cytoolz==0.9.0.1
decorator==4.4.2
dill==0.2.9
Django==1.11.9
dlib==19.17.0
docopt==0.6.2
Eel==0.10.4
en-core-web-sm==2.0.0
et-xmlfile==1.0.1
Flask==1.1.1
future==0.17.1
gevent==1.4.0
gevent-websocket==0.10.1
google-api-core==1.30.0
google-auth==1.32.0
google-cloud-texttospeech==2.4.0
googleapis-common-protos==1.53.0
googletrans==2.4.0
greenlet==0.4.15
grpcio==1.38.1
idna==2.8
imageio==2.8.0
imageio-ffmpeg==0.4.1
imutils==0.5.2
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.10.1
joblib==0.17.0
Js2Py==0.71
line-bot-sdk==1.12.1
lxml==4.4.1
MarkupSafe==1.1.1
mathparse==0.1.2
MouseInfo==0.0.4
moviepy==1.0.2
msgpack==0.5.6
msgpack-numpy==0.4.3.2
multidict==4.5.2
murmurhash==1.0.5
nltk==3.5
numpy==1.18.2
opencv-python==3.4.5.20
openpyxl==3.0.3
packaging==20.9
pafy==0.5.5
pefile==2019.4.18
pi==0.1.2
Pillow==6.2.0
pipwin==0.5.1
plac==0.9.6
preshed==2.0.1
proglog==0.1.9
proto-plus==1.19.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio @ file:///C:/Users/user/pipwin/PyAudio-0.2.11-cp37-cp37m-win32.whl
PyAutoGUI==0.9.47
pycparser==2.19
pydub==0.25.1
pygame==1.9.6
PyGetWindow==0.0.7
pyglet==1.5.3
PyInstaller==3.5
pyjsparser==2.7.1
PyMsgBox==1.0.7
PyMySQL==0.9.3
pyparsing==2.4.7
pyperclip==1.7.0
PyPrind==2.11.3
PyQt5==5.13.0
PyQt5-sip==4.19.18
pyqt5-tools==5.13.0.1.5
PyRect==0.1.4
PyScreeze==0.1.22
pySmartDL==1.3.4
python-dateutil==2.8.1
python-dotenv==0.13.0
python-vlc==3.0.7110
pytube3==9.6.4
PyTweening==1.0.3
pytz==2019.3
pywin32-ctypes==0.2.0
PyYAML==3.13
regex==2018.1.10
requests==2.22.0
requests-toolbelt==0.9.1
rsa==4.7.2
scipy==1.3.0
sip==4.19.8
six==1.16.0
snownlp==0.12.3
soupsieve==2.0
spacy==2.0.18
SpeechRecognition==3.8.1
SQLAlchemy==1.3.24
textblob==0.15.3
thinc==6.12.1
threadpool==1.3.2
toolz==0.11.1
tqdm==4.45.0
typing-extensions==3.7.4.2
tzlocal==2.1
ujson==4.0.2
urllib3==1.25.3
Werkzeug==0.15.5
whichcraft==0.6.0
wrapt==1.10.11
yarl==1.3.0
youtube-dl==2020.1.1
